Refactor: 아카이브 수정하기 컴포넌트 분리
