+++ components/Organisms/ArchiveCreate/ArchiveCreateHome.tsx
@@ -12,10 +12,10 @@ import { MapPoint } from 'assets/archive/MapPoint';
 import { Box, Button, FlexBox, RadioLabel, TextArea } from 'components/Atoms';
 import { CheckBox } from 'components/Atoms/CheckBox';
 import AddressInput from 'components/Molecules/AddressInput';
-import ArchiveFileUploadForm from 'components/Organisms/Archive/ArchiveFileUploadForm';
+import Rating from 'components/Molecules/Rating';
 import ArchiveTitle from 'components/Organisms/ArchiveCreate/ArchiveTitle';
-import Rating from 'components/Organisms/ArchiveCreate/Rating';
 import SearchTitle from 'components/Organisms/ArchiveCreate/SearchTitle';
+import ArchiveFileUploadForm from 'components/Organisms/ArchiveFileUploadForm';
 import { ALERT_MESSAGE } from 'constants/alertMessage';
 import { POPUP_NAME } from 'constants/popupName';
 import { AlertState, archiveWriteState, PopupNameState } from 'states';
