+++ components/Organisms/Detail/Navigator.tsx
@@ -53,8 +53,7 @@ export default function Navigator() {
     if (
       data?.itemInfoAdditionalInfo?.archiveLink.title === '아카이브 수정하기'
     ) {
-      // TODO (`${decode}&id=${id}`)
-      return router.push(`${decode}&exhibitionId=${id}`);
+      return router.push(decode);
     }
     return router.push({
       pathname: data?.itemInfoAdditionalInfo?.archiveLink.link,
