+++ pages/archive/update.tsx
@@ -1,13 +1,16 @@
+import { NextPage } from 'next';
 import { useRouter } from 'next/router';
 import { useEffect } from 'react';
 
-import { Box, Layout } from 'components/Atoms';
-import ArchiveHome from 'components/Organisms/Archive/ArchiveHome';
+import { Box } from 'components/Atoms';
+import NoLoginPage from 'components/Molecules/LoginPage';
+import ArchiveUpdateHome from 'components/Organisms/ArchiveUpdate/ArchiveUpdateHome';
 import { useResetArchiveByIdStateFunction } from 'states/archiveWirte';
 import theme from 'styles/theme';
+import { UserProps, useUserProps } from 'utils/authentication/useUser';
 import { useInitHeader } from 'utils/hooks/useInitHeader';
 
-export default function Archive() {
+const ArchiveUpdate: NextPage<UserProps> = ({ user }) => {
   const router = useRouter();
   const resetArchiveByIdState = useResetArchiveByIdStateFunction();
 
@@ -22,10 +25,17 @@ export default function Archive() {
     resetArchiveByIdState();
   }, [resetArchiveByIdState]);
 
+  if (!user.isLogin) {
+    return <NoLoginPage />;
+  }
+
   return (
     <Box padding="0 15px" background={theme.colors.white}>
-      <ArchiveHome />
+      <ArchiveUpdateHome />
       <Box height="60px" />
     </Box>
   );
-}
+};
+
+export const getServerSideProps = useUserProps;
+export default ArchiveUpdate;
